[AOS] feat : 오늘의 글림

+++ Glim_AOS/presentation/src/main/java/com/jeongbj/presentation/feature/home/HomeNavigation.kt
@@ -10,7 +10,7 @@ import kotlinx.serialization.Serializable
 data object HomeRoute
 
 fun NavGraphBuilder.homeNav(
-    navigateToQuoteDetail: () -> Unit,
+    navigateToQuoteDetail: (Long) -> Unit,
     navigateToBookDetail: (String) -> Unit
 ) {
     composable<HomeRoute> {
